backend/pyproject.toml
[project]
name = "booking-backend"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
  "fastapi",
  "uvicorn",
  "sqlalchemy",
  "pydantic",
  "python-dotenv"
]
